Extract files for classes in Benchmark project
