+++ Serilog.Exceptions.Benchmark/ExceptionPropertiesBag.cs
@@ -1,0 +1,60 @@
+using System;
+using System.Collections.Generic;
+using Serilog.Exceptions.Core;
+using Serilog.Exceptions.Filters;
+
+namespace Serilog.Exceptions.Benchmark
+{
+    internal class ExceptionPropertiesBag : IExceptionPropertiesBag
+    {
+        private readonly Exception exception;
+        private readonly IExceptionPropertyFilter filter;
+        private readonly Dictionary<string, object> properties = new Dictionary<string, object>();
+
+        // We keep a note on whether the results were collected to be sure that
+        // after that there are no changes. This is the application of fail-fast principle.
+        private bool resultsCollected = false;
+
+        public ExceptionPropertiesBag(Exception exception, IExceptionPropertyFilter filter = null)
+        {
+            if (exception == null)
+            {
+                throw new ArgumentNullException(nameof(exception),
+                    $"Cannot create {nameof(ExceptionPropertiesBag)} for null exception");
+            }
+
+            this.exception = exception;
+            this.filter = filter;
+        }
+
+        public IReadOnlyDictionary<string, object> GetResultDictionary()
+        {
+            this.resultsCollected = true;
+            return this.properties;
+        }
+
+        public void AddProperty(string key, object value)
+        {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key), "Cannot add exception property without a key");
+            }
+
+            if (this.resultsCollected)
+            {
+                throw new InvalidOperationException(
+                    $"Cannot add exception property '{key}' to bag, after results were already collected");
+            }
+
+            if (this.filter != null)
+            {
+                if (this.filter.ShouldPropertyBeFiltered(this.exception, key, value))
+                {
+                    return;
+                }
+            }
+
+            this.properties.Add(key, value);
+        }
+    }
+}
